% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointoss.R
\name{cointoss}
\alias{cointoss}
\title{Title}
\usage{
cointoss(n, p)
}
\arguments{
\item{n}{The number of coin tosses}

\item{p}{The probability of the coin landing on heads}
}
\value{
The outcome of each toss, with 0s for tails and 1s for heads
}
\description{
Title
}
