% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise04.R
\name{random_walk}
\alias{random_walk}
\title{Random Walk Function}
\usage{
random_walk(nsteps, step_one, step_two, prob)
}
\arguments{
\item{nsteps}{A numeric value for the number of steps of the random walk}

\item{step_one}{A numeric value for the size of the first possible step}

\item{step_two}{A numeric value for the size of the second possible step}

\item{prob}{A vector of two decimal values for the probabilities associated with each step size}
}
\value{
A tibble showing each step number and its location
}
\description{
Random Walk Function
}
